Change the `api-java` packages structure in the test module
