+++ .spine/spine_model.ser
@@ -1,3 +1,3 @@
 
-3javaclasses.exlibris.c.book.BookAggregate
-=javaclasses.exlibris.c.inventory.InventoryAggregate
+)javaclasses.exlibris.c.book.BookAggregate
+3javaclasses.exlibris.c.inventory.InventoryAggregate
