+++ api-java/src/test/java/javaclasses/exlibris/c/integrational/FlowTest.java
@@ -31,7 +31,7 @@ import io.spine.server.commandbus.CommandBus;
 import javaclasses.exlibris.BookDetailsChange;
 import javaclasses.exlibris.BoundedContexts;
 import javaclasses.exlibris.c.RemoveBook;
-import javaclasses.exlibris.c.aggregate.InventoryCommandTest;
+import javaclasses.exlibris.c.inventory.InventoryCommandTest;
 import javaclasses.exlibris.testdata.BookCommandFactory;
 import javaclasses.exlibris.testdata.BookRejectionsSubscriber;
 import javaclasses.exlibris.testdata.InventoryCommandFactory;
