+++ api-java/src/test/java/javaclasses/exlibris/c/inventory/CancelReservationCommandTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package javaclasses.exlibris.c.aggregate;
+package javaclasses.exlibris.c.inventory;
 
 import com.google.common.base.Throwables;
 import com.google.protobuf.Message;
