+++ api-java/src/test/java/javaclasses/exlibris/c/inventory/InventoryCommandTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package javaclasses.exlibris.c.aggregate;
+package javaclasses.exlibris.c.inventory;
 
 import com.google.protobuf.Message;
 import io.spine.client.TestActorRequestFactory;
@@ -26,7 +26,6 @@ import io.spine.core.CommandEnvelope;
 import io.spine.server.aggregate.AggregateCommandTest;
 import javaclasses.exlibris.BookId;
 import javaclasses.exlibris.InventoryId;
-import javaclasses.exlibris.c.inventory.InventoryAggregate;
 
 import static javaclasses.exlibris.testdata.InventoryCommandFactory.isbn62;
